+++ src/rip.h
@@ -1,4 +1,4 @@
-/* @(#) $Id: rip.h,v 1.9 1996-08-12 18:51:17 deyke Exp $ */
+/* @(#) $Id: rip.h,v 1.10 1996-08-19 16:30:14 deyke Exp $ */
 
 #ifndef _RIP_H
 #define _RIP_H
@@ -72,7 +72,7 @@ struct rip_list {
 
 /* Host format of a single entry in a RIP response packet */
 struct rip_route {
-	uint16  addr_fam;
+	uint    addr_fam;
 	int32   target;
 	int32   metric;
 };
@@ -106,13 +106,13 @@ void rip_trigger(void);
 int rip_add(int32 dest,int32 interval,int split,int us);
 int riprefadd(int32 gateway);
 int riprefdrop(int32 gateway);
-int ripreq(int32 dest,uint16 replyport);
+int ripreq(int32 dest,uint replyport);
 int rip_drop(int32 dest);
 int nbits(int32 target);
 void pullentry(struct rip_route *ep,struct mbuf **bpp);
 
 /* RIP Definition */
-extern uint16 Rip_trace;
+extern uint Rip_trace;
 extern int Rip_merge;
 extern struct rip_stat Rip_stat;
 extern struct rip_list *Rip_list;
